velar-backend/requirements.txt
fastapi>=0.115
uvicorn[standard]>=0.30
pydantic-settings>=2.0
pydantic[email]>=2.0
supabase>=2.0
python-jose[cryptography]>=3.3
asyncpg>=0.29
sqlalchemy>=2.0
